move milestones logic from properties panel to milestone selection section file
